feat(mp-djc): dedicated dashboard "Announce" button
Adds a top-level "📣 Announce" button to the admin dashboard action bar
so operators can broadcast an announcement in one click, instead of
having to open "Edit tournament details".

DRY: the ~100-line announcement composer that was inlined in
AdminEditTournament is extracted into a reusable AnnouncementComposer
component (new web-mobile/js/admin_announcement.jsx), plus an
AnnouncementModal wrapper using the existing .modal-backdrop pattern.
The composer is now reused by both the dashboard modal and the
Edit-details page (kept as a fallback). Backend endpoint is unchanged
(POST /api/tournament/announce).

The modal uses .modal--lg (720px) — the width the composer was designed
for on the Edit-details card; at the default 460px the message-hint
row's space-between flex wrapped the char counter mid-sentence.

The pure helpers isSendAnnouncementDisabled / sendAnnouncementLabel
move to the new file alongside the component they drive; their unit
tests move to admin_announcement.test.jsx.
