+++ web-mobile/index.html
@@ -66,6 +66,7 @@
     <script type="module" src="/dist/admin_pools.js?v=5"></script>
     <script type="module" src="/dist/admin_schedule.js?v=5"></script>
     <script type="module" src="/dist/admin_competition.js?v=5"></script>
+    <script type="module" src="/dist/admin_announcement.js?v=5"></script>
     <script type="module" src="/dist/admin_setup.js?v=5"></script>
     <script type="module" src="/dist/admin.js?v=5"></script>
     <script type="module" src="/dist/app.js?v=5"></script>
